+++ src/app/core/database/models/create-goal-payload.ts
@@ -1,0 +1,8 @@
+import { GoalTargetValueUnit } from '@core/database/models/goal-target-value-unit';
+
+export interface CreateGoalPayload {
+	name: string;
+	targetValue: number;
+	unit: GoalTargetValueUnit;
+	domainId: number;
+}