+++ src/app/features/account/services/delete-account/delete-account.service.ts
@@ -1,12 +1,12 @@
 import { inject, Injectable } from '@angular/core';
 import { defer, from, Observable } from 'rxjs';
-import { DeleteAllService } from '@core/database/services/delete-all/delete-all.service';
+import { ClearDatabaseService } from '@core/database/services/clear-database/clear-database.service';
 
 @Injectable({ providedIn: 'root' })
 export class DeleteAccountService {
-	private readonly deleteAllService = inject(DeleteAllService);
+	private readonly clearDatabaseService = inject(ClearDatabaseService);
 
 	deleteAccount(): Observable<void> {
-		return defer(() => from(this.deleteAllService.deleteAll()));
+		return defer(() => from(this.clearDatabaseService.clearDatabase()));
 	}
 }