+++ src/app/pages/domains/:id/goals/domain-goals-page.ts
@@ -1,35 +1,26 @@
 import { Component, DestroyRef, inject, OnInit, signal } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { ActivatedRoute } from '@angular/router';
 import { finalize } from 'rxjs';
 import { MessageService } from 'primeng/api';
 import { ButtonModule } from 'primeng/button';
-import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
-import { DEFAULT_ERROR_MESSAGE } from '@core/constants/messages';
 import { Domain } from '@core/database/models/domain';
+import { DEFAULT_ERROR_MESSAGE } from '@core/constants/messages';
 import { FetchDomainService } from '@features/domain/services/fetch-domain/fetch-domain.service';
-import { DeleteDomainDialog } from '@features/domain/components/delete-domain-dialog/delete-domain-dialog';
-import { UpdateDomainDialog } from '@features/domain/components/update-domain-dialog/update-domain-dialog';
 
 @Component({
-	selector: 'orb-domain-page',
+	selector: 'orb-domain-goals',
 	imports: [ButtonModule],
-	templateUrl: './domain-page.html',
+	templateUrl: './domain-goals-page.html',
 	host: {
 		class: 'flex-1 px-3',
 	},
-	providers: [DialogService],
 })
-export class DomainPage implements OnInit {
+export class DomainGoalsPage implements OnInit {
 	private readonly activatedRoute = inject(ActivatedRoute);
 	private readonly fetchDomainService = inject(FetchDomainService);
 	private readonly messageService = inject(MessageService);
-	private readonly router = inject(Router);
-	private readonly dialogService = inject(DialogService);
 	private readonly destroyRef = inject(DestroyRef);
-
-	ref1?: DynamicDialogRef | null;
-	ref2?: DynamicDialogRef | null;
 
 	protected readonly domainId = signal<number | null>(null);
 	protected readonly domain = signal<Domain | null>(null);
@@ -67,50 +58,10 @@ export class DomainPage implements OnInit {
 					this.messageService.add({
 						key: 'main',
 						severity: 'error',
-						summary: 'Fetch Failure',
+						summary: 'Domain Fetch Failure',
 						detail: err?.message ?? DEFAULT_ERROR_MESSAGE,
 					});
 				},
 			});
-	}
-
-	protected updateDomain() {
-		this.ref1 = this.dialogService.open(UpdateDomainDialog, {
-			header: 'Update your domain',
-			inputValues: { domain: this.domain() },
-			width: '50vw',
-			modal: true,
-			closable: true,
-			breakpoints: {
-				'960px': '75vw',
-				'640px': '90vw',
-			},
-		});
-
-		this.ref1?.onClose.subscribe((isUpdated) => {
-			if (typeof isUpdated === 'boolean' && isUpdated) {
-				this.fetchDomainById();
-			}
-		});
-	}
-
-	protected deleteDomain() {
-		this.ref2 = this.dialogService.open(DeleteDomainDialog, {
-			header: 'Delete your domain',
-			inputValues: { domain: this.domain() },
-			width: '50vw',
-			modal: true,
-			closable: true,
-			breakpoints: {
-				'960px': '75vw',
-				'640px': '90vw',
-			},
-		});
-
-		this.ref2?.onClose.subscribe((isDeleted: unknown) => {
-			if (typeof isDeleted === 'boolean' && isDeleted) {
-				this.router.navigate(['/domains']);
-			}
-		});
 	}
 }