+++ src/app/pages/domains/domains-page.ts
@@ -86,6 +86,7 @@ export class DomainsPage implements OnInit {
 			width: '50vw',
 			modal: true,
 			closable: true,
+			closeOnEscape: true,
 			breakpoints: {
 				'960px': '75vw',
 				'640px': '90vw',