+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 - Added the menu.
 - Added account deletion and links to the source.
 - Added account update.
+- Added domains.
+- Added domain goals.
 
 ### Changed